requirements.txt
treys
httpx
pyyaml
rich
fastapi
uvicorn[standard]
